+++ tests/test_db.py
@@ -3,13 +3,45 @@
 
 import unittest
 from types import SimpleNamespace
-
+import asyncio
 
 import asynctest
 import asyncpg
-
+import aiohttp.web
 
 from swift_x_account_sharing.db import DBConn
+from swift_x_account_sharing.db import handle_dropped_connection
+
+
+class HandleDroppedTestClass(asynctest.TestCase):
+    """Test class for dropped connection handling."""
+
+    def setUp(self):
+        """Set up necessary mocks."""
+        self.mock_request = SimpleNamespace(**{
+            "app": {
+                "db_conn": SimpleNamespace(**{
+                    "erase": unittest.mock.Mock(),
+                    "open": asynctest.mock.Mock(),
+                }),
+            },
+        })
+
+        self.ensure_future_mock = unittest.mock.MagicMock(
+            asyncio.ensure_future
+        )
+        self.patch_asyncio_ensure_future = unittest.mock.patch(
+            "swift_x_account_sharing.db.asyncio.ensure_future",
+            new=self.ensure_future_mock
+        )
+
+    async def test_handle_dropped_connection(self):
+        """Test handling dropped connection."""
+        with self.patch_asyncio_ensure_future:
+            with self.assertRaises(aiohttp.web.HTTPServiceUnavailable):
+                await handle_dropped_connection(self.mock_request)
+            self.mock_request.app["db_conn"].erase.assert_called_once()
+            self.mock_request.app["db_conn"].open.assert_called_once()
 
 
 class DBConnTestClass(asynctest.TestCase):
@@ -132,6 +164,7 @@ class DBMethodTestCase(asynctest.TestCase):
                     "container": "test-container",
                     "container_owner": "test-owner",
                     "recipient": "test-receiver",
+                    "sharingdate": "testdate",
                     "address": "test-address",
                     "r_read": True,
                     "r_write": True,
@@ -142,6 +175,7 @@ class DBMethodTestCase(asynctest.TestCase):
                     "r_read": True,
                     "r_write": True,
                     "container": "test-container",
+                    "sharingdate": "testdate",
                     "container_owner": "test-owner",
                     "recipient": "test-recipient",
                     "address": "test-address",
